requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
django-jazzmin==2.6.0
django-phonenumber-field==7.3.0
Faker==22.5.1
idna==3.6
phonenumbers==8.13.29
pillow==10.2.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0